[refact/#113]: 컴포넌트 분리
