+++ src/app/mission/list/page.jsx
@@ -8,7 +8,6 @@ export default function Page() {
       <div className="flex flex-col h-screen overflow-hidden">
         <Header />
         <div className="p-4 flex-1 overflow-hidden flex flex-col">
-          <h2 className="text-lg mb-4">아이 선택 및 필터링 여하고 부분</h2>
           <div className="flex-1 overflow-y-auto custom-scrollbar">
             <MissionList />
           </div>
